+++ NEWS.md
@@ -2,7 +2,14 @@
 
 ## Improvements
 
-- New AI agent capabilities
+- New AI Agent Capabilities
+  - `iterate_forecast()` can use LLM's to find the optimal combination of data and inputs to create the most accurate forecast. 
+  - `update_forecast()` can take previously trained models from `iterate_forecast()` to create forecasts on new data fast.
+  - Helper functions `set_project_info()` and `set_agent_info` to assist in iterating and updating forecasts. 
+  
+## Breaking Changes
+
+- `experiment_name` within `set_run_info()` has been changed to `project_name` to comply with new AI agent capabilities. 
 
 # finnts 0.6.0